+++ Poker/Node/ShowdownNode.cs
@@ -48,7 +48,7 @@ namespace Poker
             uint PocketValue1, PocketValue2;
             for (int p1 = 0; p1 < Pocket.N; p1++)
             {
-                //if (double.IsNaN(PocketP[p1])) continue;
+                //if (decimal.IsNaN(PocketP[p1])) continue;
                 if (!MyCommunity.AvailablePocket[p1]) continue;
                 PocketValue1 = PocketValue[p1];
                 Assert.That(PocketValue1 < uint.MaxValue);
@@ -58,10 +58,10 @@ namespace Poker
                 UpdateOnExclusion(PocketP, UpdatedP, p1);
                 OpCount++;
 
-                double ShowdownEV = 0;
+                decimal ShowdownEV = 0;
                 for (int p2 = 0; p2 < Pocket.N; p2++)
                 {
-                    //if (double.IsNaN(UpdatedP[p2])) continue;
+                    //if (decimal.IsNaN(UpdatedP[p2])) continue;
                     if (!MyCommunity.AvailablePocket[p2]) continue;
                     PocketValue2 = PocketValue[p2];
                     Assert.That(PocketValue2 < uint.MaxValue);
@@ -80,7 +80,7 @@ namespace Poker
             /* Assymptotically optimal implementation. O(N^2) */
             RiverCommunity River = (RiverCommunity)MyCommunity;
             Optimize.ProbabilityPrecomputation(PocketP, MyCommunity);
-            double Correction;
+            decimal Correction;
             int _p1;
 
             // For each pocket we might have, calculate the chance to win.
@@ -90,7 +90,7 @@ namespace Poker
             while (_p1 < Pocket.N)
             {
                 int p1 = River.SortedPockets[_p1];
-                //if (double.IsNaN(PocketP[p1])) { _p1++; continue; }
+                //if (decimal.IsNaN(PocketP[p1])) { _p1++; continue; }
                 if (!MyCommunity.AvailablePocket[p1]) { _p1++; continue; }
 
                 // Find next highest pocket
@@ -103,13 +103,13 @@ namespace Poker
                 for (int EqualValuedPocket = _p1; EqualValuedPocket < NextHighest; EqualValuedPocket++)
                 {
                     int p = River.SortedPockets[EqualValuedPocket];
-                    //if (double.IsNaN(PocketP[p])) continue;
+                    //if (decimal.IsNaN(PocketP[p])) continue;
                     if (!MyCommunity.AvailablePocket[p]) continue;
 
                     var pocket1 = Pocket.Pockets[p];
                     int c1 = pocket1.Cards[0], c2 = pocket1.Cards[1];
 
-                    double ChanceToWin =
+                    decimal ChanceToWin =
                     Optimize.SummedChance -
                         Optimize.SummedChance_OneCardFixed[c1] -
                         Optimize.SummedChance_OneCardFixed[c2];
@@ -128,13 +128,13 @@ namespace Poker
                 for (int EqualValuedPocket = _p1; EqualValuedPocket < NextHighest; EqualValuedPocket++)
                 {
                     int p = River.SortedPockets[EqualValuedPocket];
-                    //if (double.IsNaN(PocketP[p])) continue;
+                    //if (decimal.IsNaN(PocketP[p])) continue;
                     if (!MyCommunity.AvailablePocket[p]) continue;
 
                     var pocket1 = Pocket.Pockets[p];
                     int c1 = pocket1.Cards[0], c2 = pocket1.Cards[1]; 
                     
-                    double P = PocketP[p];
+                    decimal P = PocketP[p];
 
                     Optimize.SummedChance += P;
                     Optimize.SummedChance_OneCardFixed[c1] += P;
@@ -151,7 +151,7 @@ namespace Poker
             while (_p1 > 0)
             {
                 int p1 = River.SortedPockets[_p1];
-                //if (double.IsNaN(PocketP[p1])) { _p1--; continue; }
+                //if (decimal.IsNaN(PocketP[p1])) { _p1--; continue; }
                 if (!MyCommunity.AvailablePocket[p1]) { _p1--; continue; }
 
                 // Find next highest pocket
@@ -164,13 +164,13 @@ namespace Poker
                 for (int EqualValuedPocket = _p1; EqualValuedPocket > NextLowest; EqualValuedPocket--)
                 {
                     int p = River.SortedPockets[EqualValuedPocket];
-                    //if (double.IsNaN(PocketP[p])) continue;
+                    //if (decimal.IsNaN(PocketP[p])) continue;
                     if (!MyCommunity.AvailablePocket[p]) continue;
 
                     var pocket1 = Pocket.Pockets[p];
                     int c1 = pocket1.Cards[0], c2 = pocket1.Cards[1];
 
-                    double ChanceToLose =
+                    decimal ChanceToLose =
                     Optimize.SummedChance -
                         Optimize.SummedChance_OneCardFixed[c1] -
                         Optimize.SummedChance_OneCardFixed[c2];
@@ -189,13 +189,13 @@ namespace Poker
                 for (int EqualValuedPocket = _p1; EqualValuedPocket > NextLowest; EqualValuedPocket--)
                 {
                     int p = River.SortedPockets[EqualValuedPocket];
-                    //if (double.IsNaN(PocketP[p])) continue;
+                    //if (decimal.IsNaN(PocketP[p])) continue;
                     if (!MyCommunity.AvailablePocket[p]) continue;
 
                     var pocket1 = Pocket.Pockets[p];
                     int c1 = pocket1.Cards[0], c2 = pocket1.Cards[1];
 
-                    double P = PocketP[p];
+                    decimal P = PocketP[p];
 
                     Optimize.SummedChance += P;
                     Optimize.SummedChance_OneCardFixed[c1] += P;
@@ -207,7 +207,7 @@ namespace Poker
 #endif
         }
 
-        public override double _Simulate(Var S1, Var S2, int p1, int p2, ref int[] BranchIndex, int IndexOffset)
+        public override decimal _Simulate(Var S1, Var S2, int p1, int p2, ref int[] BranchIndex, int IndexOffset)
         {
             uint PocketValue1 = PocketValue[p1];
             uint PocketValue2 = PocketValue[p2];
